Successfully Posting
Able to post to both source and target tumblr blogs.
Later, will move the definition of source and target to the command line, but for now the .env is more convenient.
Also laid out skeleton for future functions. Used same commenting system as andrewjkerr's tumblr-mass-reblog-by-tag, which has been very helpful for grasping how to work with this API.